backend/requirements.txt
Django==4.1.9
djangorestframework==3.14.0
mysqlclient==2.1.1
django-mysql==4.9.0
django-cors-headers==4.0.0
djoser==2.2.0